Fix cleanup of L2 hosts
The cleanup scripted missed cleaning up hosts that are attached
to a router with a L2 network.

Additionally, fixes a logic bug in determining if a
L2 router should also have a host attached.

router_config.txt
GENE	N/A	L2-UNIV:thomahol/d_host
Correctly starts a host container using the thomahol/d_host image

Where as
GENE    N/A     L2-UNIV
Attempts to start a host container using the L2-UNIV image
which will never exist.

The 'cut' command used to split at the ':' returns
the full string when the delimiter is missing.
The fix is to add the '-s' (--only-delimited) option to prevent
the original string being returned in this case.